Create separate class for testing Habit reorder
It otherwise clutters the HabitTest class with methods only related to a
single unrelated test.
